+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Require `chain_id` parameter for registering subscription option.
 - Import posts from outbox in chronological order.
+- Add `created_at` parameter to identity claim API response.
 
 ### Fixed